+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- Add cli reset command (gorarakelyan)
 - Add nested module tracking support to aim sdk (gorarakelyan)
 - Add code difference tracking to aim sdk (gorarakelyan)
 - Update aim push command to send commits (gorarakelyan)